Style: 폰트 컴포넌트수정 스플레쉬 연결
