+++ src/pages/Welcome/Welcome.jsx
@@ -40,6 +40,7 @@ const KakaoLogin = styled.button`
   border-radius: 44px;
   padding: 13px 87px;
   background: url(${KakaoLogo}) no-repeat 12px 9px;
+  /* white-space: nowrap; */
 `;
 const GoogleLogin = styled.button`
   width: 322px;
@@ -58,6 +59,7 @@ const FacebookLogin = styled.button`
   border-radius: 44px;
   padding: 13px 87px;
   background: url(${FacebookLogo}) no-repeat 12px 9px;
+  /* white-space: nowrap; */
 `;
 const LoginJoinBox = styled.div`
   display: flex;
